+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/rest_api/models/enums.py
@@ -233,6 +233,11 @@ class ModifyOrderPriceMatchEnum(Enum):
     QUEUE_20 = "QUEUE_20"
 
 
+class ModifyOrderReduceOnlyEnum(Enum):
+    TRUE = "true"
+    FALSE = "false"
+
+
 class NewAlgoOrderAlgoTypeEnum(Enum):
     CONDITIONAL = "CONDITIONAL"