+++ clients/derivatives_trading_usds_futures/tests/unit/websocket_api/test_trade_api_ws_api.py
@@ -29,6 +29,11 @@ from binance_sdk_derivatives_trading_usds_futures.websocket_api.models import (
 
 
 from binance_sdk_derivatives_trading_usds_futures.websocket_api.models import (
+    ModifyOrderReduceOnlyEnum,
+)
+
+
+from binance_sdk_derivatives_trading_usds_futures.websocket_api.models import (
     NewAlgoOrderAlgoTypeEnum,
 )
 
@@ -578,6 +583,7 @@ class TestWebSocketTradeApi:
             "orig_client_order_id": "1",
             "price_match": ModifyOrderPriceMatchEnum["OPPONENT"].value,
             "modify_id": 1,
+            "reduce_only": ModifyOrderReduceOnlyEnum["TRUE"].value,
             "recv_window": 5000,
         }
 
@@ -650,6 +656,7 @@ class TestWebSocketTradeApi:
         assert params["orig_client_order_id"] == "1"
         assert params["price_match"] == ModifyOrderPriceMatchEnum["OPPONENT"].value
         assert params["modify_id"] == 1
+        assert params["reduce_only"] == ModifyOrderReduceOnlyEnum["TRUE"].value
         assert params["recv_window"] == 5000
 
         assert result is not None
@@ -674,6 +681,7 @@ class TestWebSocketTradeApi:
             "orig_client_order_id": "1",
             "price_match": ModifyOrderPriceMatchEnum["OPPONENT"].value,
             "modify_id": 1,
+            "reduce_only": ModifyOrderReduceOnlyEnum["TRUE"].value,
             "recv_window": 5000,
         }
         params["symbol"] = None
@@ -695,6 +703,7 @@ class TestWebSocketTradeApi:
             "orig_client_order_id": "1",
             "price_match": ModifyOrderPriceMatchEnum["OPPONENT"].value,
             "modify_id": 1,
+            "reduce_only": ModifyOrderReduceOnlyEnum["TRUE"].value,
             "recv_window": 5000,
         }
         params["side"] = None
@@ -716,6 +725,7 @@ class TestWebSocketTradeApi:
             "orig_client_order_id": "1",
             "price_match": ModifyOrderPriceMatchEnum["OPPONENT"].value,
             "modify_id": 1,
+            "reduce_only": ModifyOrderReduceOnlyEnum["TRUE"].value,
             "recv_window": 5000,
         }
         params["quantity"] = None
@@ -739,6 +749,7 @@ class TestWebSocketTradeApi:
             "orig_client_order_id": "1",
             "price_match": ModifyOrderPriceMatchEnum["OPPONENT"].value,
             "modify_id": 1,
+            "reduce_only": ModifyOrderReduceOnlyEnum["TRUE"].value,
             "recv_window": 5000,
         }
         params["price"] = None